Add initial guide for test data requests
Those of use new to the team are already starting to come across various examples and 'tid-bits' from the previous development team that cover API requests used to teardown and then seed the DB with data.

Much as it appears to be a rather clunky and rigid solution, inline with trying to capture and record these things in one place this change adds an initial guide.

It'll give an overview based on what we know and then example requests.

Note - Though these endpoints are in the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) we're keeping them in their own `test_data.md` guide to differentiate them from _actual_ service endpoint examples.