Add details on background jobs
Various background jobs exist within the service which are run by various apps. This is a guide to what jobs are running where, and when. As we learn more we'll also start recording what they are doing.

But for now this is a simple list by service, type and if we know it, times a job runs.