Guide to ALCS supplementary bill run process
The dev team have been tasked with developing the SROC version of the supplementary WRLS bill run process. Before we do that we need to understand the ALCS (pre-SROC) one already in the system.

This guide (draft to begin with) is our notes on hwta we found and what we believe the current code is doing (to the best of our understanding).