Update to match SROC CHA API
We intend to add a guide on scheduling and what to do on release day in subsequent changes. So, we add a 'releasing' folder ready to take all these guides, the first of which is 'sign-off'.

We always add a root README.md to our folders. So, again we use the existing SROC CHA API one as a basis and update it to be relevant to WATER.