Add start of releasing docs - sign-off (#57)
https://eaflood.atlassian.net/browse/WATER-3720

Preparing a 'release' and arranging getting it into `production` can be a convoluted process. Often someone on the team becomes 'the person' who deals with it. The problem is should that person be unavailable delivery teams can become blocked or releases get delayed whilst they work things out in their absence.

To avoid that we document the process and update it should anything change. It means anyone from the tech team _should_ be able to conduct a release and no one person becomes a liability.

The first step in the release process is signing off the 'the release'. This change adds the new section to our docs and the first guide; sign off.