Add next releasing doc - release
https://eaflood.atlassian.net/browse/WATER-3720

This adds a guide on the last step of our release to production process; release.

The tech team are responsible for supporting each release to production. In nearly all cases for the devs this just means being on hand and letting folks know when it's complete. web-ops will push the button and the QA will smoke test the service.

But just in case something goes wrong, the guide also covers what to do in the event of an issue.