Guide to redis usage across the service (#47)
This originated when trying to sort out the Redis configuration in the [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui/pull/2190) by bringing it in line with how we do things in the water-abstraction-service.

All the projects that use Redis specify the [DB number](https://redis.io/commands/select/). The default is 0, but if you change that in a connection config you'll create and then use a separate DB within the same Redis instance. The WRLS system uses this to allocate different DB numbers across the apps and in the case of water-abstraction-service, different DB numbers for live and testing.

Problem is, it's all still on the same Redis instance which means we need to keep track of who's using what, else it defeats the point! We should also point out what the Redis docs say

> In practical terms, Redis databases should be used to separate different keys belonging to the same application (if needed), and **not** to use a single Redis instance for multiple unrelated applications.

Oh boy! 😩🤦

So, the primary purpose of this guide is to track these numbers in one place. But it'll also give us somewhere to hang any other pertinent Redis info.