Add section and guides for RfC's
https://eaflood.atlassian.net/browse/WATER-3720

Though there are multiple folks within the team who have the right access in MyIT to raise RfC's not everyone is confident or aware of how to complete the form. On a previous teams, we created guides to help with this.

- [Ruby Services Team](https://github.com/DEFRA/ruby-services-team/blob/main/request_for_change.md)
- [SROC Service Team](https://github.com/DEFRA/sroc-service-team/blob/main/rfc/README.md)

This change adds similar guides for completing RfC's for our service, including an example.