Add section and guides for RfC's (#63)
https://eaflood.atlassian.net/browse/WATER-3720

Though there are multiple folks within the team who have the right access in MyIT to raise RfCs not everyone is confident or aware of how to complete the form. On previous teams, we created guides to help with this.

- [Ruby Services Team](https://github.com/DEFRA/ruby-services-team/blob/main/request_for_change.md)
- [SROC Service Team](https://github.com/DEFRA/sroc-service-team/blob/main/rfc/README.md)

This change adds similar guides for completing RfCs for our service, including examples.