Add initial code on coding conventions (#68)
Now [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) is starting to take shape there are some 'unwritten' coding conventions we're applying. So, about time they became 'written'! 😁

This adds a guide which documents the initial ones we've noted. It is _strongly_ encouraged that the team actively contribute to this guide as and when they think a new convention has become known.