Add ad-hoc queries section (#75)
From time to time we are asked to create and run ad-hoc queries to extract data from the service. There are also those we create for ourselves or those we discover in the existing service.

We think there is some value in being able to reference these queries when creating new ones in the future. So, we have created a new section in our teams' repo.