Ad-hoc query for lic. with charge vers. in 2022
https://eaflood.atlassian.net/browse/WATER-3901

This is the first of 2 queries requested by the business to support the need to accrue due to the late delivery of SROC supplementary billing.

The request for the first report is

> [..] all licences (Live and Dead) with a confirmed charge version created on or after 1/4/2022 [..]

The data requested was

- Licence Number
- Licence Status
- Licence Type (full or transfer)
- Effective end date if not live (revoked, lapsed or expired date)
- Detail 1 line per charge category (and or purpose) containing all fields from the the charge version
- Include additional field to show if an agreement is in effect
- Agreement type in effect (eg 2PT)
- Include Y/N field to show if the Annual Bill has been run for the licence

What we're outputting is not quite the same, but we feel it's better to give the raw data which some of these properties can then be derived from.