Updated following initial feedback
> Note - any names used are what things are referred to in the DB

The key changes are:

- if the expired date is not yet hit then the licence remains current
- revoked and lapsed end dates now take precedence over expired
- all charge versions for a licence that start 2022-04-01 need to be included, not just the current one
- extract specific fields from the `charge_elements` table
- extract additional charges and adjustment JSONB fields as distinct columns

There was some confusion as to why lines appeared to be repeating. It was was because we were asked to include 'all' charge information, but only the charge version fields. The hierarchy in the DB is

- licence
  - charge version
    - charge element
      - charge purpose

So, you could have a licence like this
- licence
  - charge version
    - charge element
      - charge purpose
      - charge purpose
    - charge element
      - charge purpose
  - charge version
    - charge element
      - charge purpose
      - charge purpose

The number of charge purposes determines how many times a licence appears in the results as it's the lowest level in the hierarchy. So, we have tried to make this clearer by grouping the data in this hierarchical structure dropping the ID column before each group. We have also included anything that looked pertinent from the charge purpose record, if only to show why a 'charge reference' appears to have multiple lines.