Ad-hoc query for lic. without charge vers. in 2022 (#77)
https://eaflood.atlassian.net/browse/WATER-3901

This is the first of 2 queries requested by the business to support the need to accrue due to the late delivery of SROC supplementary billing.

The request for the first report is

> [..] all licences (Live and dead) where no confirmed charge version exists with an effective date on or after 1/4/22 [..]

The data requested was

- Licence Number
- Licence Status
- effective start date
- Effective end date if not live (revoked, lapsed or expired date)
- Detail 1 line per charge category (and or purpose) containing all fields from the charge version
- Include additional field to show if an agreement is in effect
- Agreement type in effect (eg 2PT)
- Y/N Indicator to show if Licence is in Workflow
- Workflow status (to set up, in review)

What we eventually built is not quite the same. We iterated the initial query a few times after feedback. But this was the genesis of the query.