Add licences with primary users query
https://eaflood.atlassian.net/browse/WATER-4047

To help monitor KPI's for the service the data team would like a second extract of licences in the service that have a primary user. This means an external user has come to the service and registered themselves as the primary user for a licence.

We had done this previously in January 2023 (see WATER-4047) but at that time we were not recording ad-hoc queries in this repo.

This change adds the query we originally used for future use.