Add query all billing accounts and their addresses
https://eaflood.atlassian.net/browse/WATER-4054

Billing & Data team have requested an ad-hoc report of all billing accounts and the addresses for them. They are specifically looking to identify those with a foreign address as SSCL have to be informed so they can deal with the invoices in SOP.

The B&D team have agreed to do the identifying as long as we can provide the raw data.

So, this query pulls all `crm_v2.invoice_accounts` and their linked addresses. This can then be exported to CSV.