Add last billed ad-hoc query
https://eaflood.atlassian.net/browse/WATER-4136

The B&D team have been wanting a 'billed up to' date against a charge version to help them manage billing. It was something they had in NALD but has been lost in WRLS.

We think there might be a way to see when a charge version was last billed using the existing data, specifically the fact that the `water.billing_transactions` stores a reference to the `water.charge_elements` it was based on.

This change covers 2 queries. The first is the main one that determines for each `water.charge_version` what was the last annual, supplementary and two-part tariff bill run it was part of. The second can be used to help verify the results. Given a licence reference it will get all the `water.billing_batches` linked to it. Any bill runs listed in the first should be in the verification script.