Add first bill run after chg ver ad-hoc query
https://eaflood.atlassian.net/browse/WATER-4136

The B&D team wants a 'billed up to' date against a charge version to help them manage billing. It was something they had in NALD but has been lost in WRLS.

We think there might be a way to see what was the first bill run created after a charge version is created by comparing their `date_created` fields.

This change adds the query we generated. It is able to report on annual and supplementary bill runs separately for each charge version rather than report a simple 'first possible bill run'.