Add guides and info for new starters
https://eaflood.atlassian.net/browse/WATER-4243
https://eaflood.atlassian.net/browse/WATER-4244

Our team is expanding! Which means we have new folks we want to get comfortable and happy in the team as quickly as possible. The last time we went through this we relied a little too much on person-2-person updates. The existing team found they couldn't make themselves available as much as they would have liked. So, a 'dump' of info isn't going to be immediately retained by someone on their first day!

So, this is the first iteration of trying to get some of that knowledge documented so new starters always have something they can refer to. It also means should we be hit by a range of live issues just when the new folks start, there is something they can get on with.

This is the first iteration and as always with documentation, is being done at the last minute! But it hopefully give us a foundation to build on and improve.