Remove 'Do this first' from welcome guide
https://eaflood.atlassian.net/browse/WATER-4243
https://eaflood.atlassian.net/browse/WATER-4244

We hastily threw the [welcome guide together](https://github.com/DEFRA/water-abstraction-team/pull/108) so we had something for some new starters to our team. When we wrote it we thought individuals themselves had to summit the `CCOE Azure/AWS Non-Production Service Request` to get access to non-prod environments and things like Jenkins and GitLab.

But having walked them through it we saw we can summit these requests on their behalf. Just as soon as we know they have a Defra email address we can get the requests in. It makes for a much slicker process so there is no need to have this call to action in the guide.

This change removes it.