Alter 'Do this first' section in welcome guide (#109)
https://eaflood.atlassian.net/browse/WATER-4243
https://eaflood.atlassian.net/browse/WATER-4244

We hastily threw the [welcome guide together](https://github.com/DEFRA/water-abstraction-team/pull/108) so we had something for some new starters to our team. When we wrote it we thought individuals themselves had to summit the `CCOE Azure/AWS Non-Production Service Request` to get access to non-prod environments and things like Jenkins and GitLab.

But having walked them through it we saw we can submit these requests on their behalf. Just as soon as we know they have a Defra email address we can get the requests in. It makes for a much slicker process.

So, we're altering this section so the call to action is on checking the tech lead has submitted the request rather than asking them to do it themselves. It means we can leave the details in so the tech lead has a chance of remembering what it is they are supposed to be doing!