Add new adhoc scripts section to project
https://eaflood.atlassian.net/browse/WATER-4356

Working on a [feature](https://eaflood.atlassian.net/browse/WATER-4345) recently we needed a way of extracting and importing SROC charge information into a local environment. We were working on annual billing so we needed _a lot_ of data!

We came up with an adhoc script to do most of the work as we believe this is something we'll very infrequently need to do. So, we didn't want to lose what we'd learned but we also didn't want to invest time making it a robust, `production` ready solution.

Hence a new **adhoc** section in our team guides & docs repo!

This creates the folder and adds our first adhoc script.