Add coding convention for unit test descriptive text (#119)
There has been some inconsistency in how words are being emphasised in the `describe()` or `it()` blocks of a unit test.

Following a discussion on our daily dev call we agreed on a standard where the descriptive text would always be wrapped in single quotes. If there is a word(s) within that text that we want to emphasise then this would be wrapped in double quotes.

This PR adds this convention to our coding conventions document.