Update welcome document to list Discord not Jitsi (#121)
https://github.com/DEFRA/water-abstraction-team/issues/99

We just spotted that the `welcome.md` document lists [Jitsi Meet](https://meet.jit.si/) as our video conferencing tool. We actually use [Discord](https://discord.com/) and have been doing so for a number of months.  We found **Jitsi** a little too flaky for our tastes. **Discord** was suggested instead, and so far it has proven more robust, and we haven't encountered any limitations on use.

This change corrects the guide.