Temporary session storage cleanup job https://eaflood.atlassian.net/browse/WATER-4281
As part of the work on returns, we have had to build in support for temporary session storage into the `water-abstraction-system`. This is what allows us to track data across pages as part of a journey until it’s ready to be properly persisted.

A job has now been created that can be triggered through an API call that deletes any record in sessions where the created on date is more than 24 hours ago.

This PR is going to add details of that job to our documentation.