Update job notes to reflect current status
https://eaflood.atlassian.net/browse/WATER-4670
https://eaflood.atlassian.net/browse/WATER-4728

At the end of 2024, we had issues with the import. We traced it down to changes we made to start migrating away from it. When these were reverted all was well again.

> N.B. The plan to rebuild the legacy import is now on hold as NALD is no longer being replaced by ReSP

As part of that process though, we did a deep dive on the import and learnt a lot about how long the various jobs were taking. With this new information we rescheduled when things happen, but it looks like we never updated our notes.

We've also just amended the **NALD Import** and **Licence Import** jobs. So, now seems a good time to update our notes about the jobs.