Add query to verify licences 4 quarterly returns (#133)
https://eaflood.atlassian.net/browse/WATER-4880

To switch a licence to use quarterly return submissions, you have to create a new return version with the quarterly flag set. Rather than make our users do this manually, we intended to perform a one-off migration.

To do that, though, we need to know we're adding the new return version to the correct licences.

This change records an ad-hoc we created to find 'water undertaker' licences that 'quarterly-returns' applies to. It outputs when they end and whether they have an existing return version we can copy from. It also flags those we believe we need to create quarterly return versions for, and those we'll have a problem with if we try.

The output of this will be sent to the business to be verified, and used to correct the existing data ahead of the migration.