Add guide on restoring prod DB to pre-prod (#137)
There are times we need to restore the `production` database into our `pre-prodcution` environment, for example, UAT or investigating an issue.

Our web-ops team have provided a Jenkins job that restores the last `prod` backup using [pg_restore](https://www.postgresql.org/docs/current/app-pgrestore.html).

We have found that issues can arise if the apps are running and the DB is not cleared beforehand. So, we have a process that helps avoid this.

This change adds a guide to the process for use by the team.