Update Licences with a primary user query
https://eaflood.atlassian.net/browse/WATER-4969

Back in March we were asked to re-run the **Licences with a primary user** report but with some changes.

The business were looking to do a 'comms drop' to our registered users, which might include a questionnaire. They wanted to be able to group the information by area and industry hence the need for a tweaked version.

After seeing the results, they asked if this could be used in place of the existing query going forward.

We've just been asked to run it again, only we ran the original. That's when they reminded us they wanted the new version.

So, making this change to avoid a similar mistake in the future.