Update query for billing account addresses
https://eaflood.atlassian.net/browse/WATER-5326

Billing & data asked for a refresh of the extract in WATER-5313. But what they really wanted was just the 'current' address for each billing account.

So, we amended the query to select only these from `crm_v2.invoice_account_addresses`. The extra start and end date fields were just to help them confirm we'd got the 'right' address records.